More warning fixes

+++ cocos/2d/renderer/QuadCommand.cpp
@@ -23,7 +23,7 @@ QuadCommand::QuadCommand()
     _quad = nullptr;
 }
 
-void QuadCommand::init(int viewport, int32_t depth, GLuint textureID, GLProgram* shader, BlendFunc blendType, V3F_C4B_T2F_Quad* quad, int quadCount, const kmMat4 &mv)
+void QuadCommand::init(int viewport, int32_t depth, GLuint textureID, GLProgram* shader, BlendFunc blendType, V3F_C4B_T2F_Quad* quad, ssize_t quadCount, const kmMat4 &mv)
 {
     _viewport = viewport;
     _depth = depth;
